Move sketch to an appropriately named folder
The Arduino IDE requires that a sketch be located in a folder of the same name. When opening a file that does not meet this requirement the Arduino IDE presents a dialog:

The file "lighthub.ino" needs to be inside a sketch folder named "lighthub".
Create this folder,  move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file lighthub.ino to the new folder, leaving behind the other sketch files. This causes compilation of the sketch to fail "fatal error: options.h: No such file or directory".

Although the repository name does match the sketch name, it is common for people to use GitHub's "Clone or Download > Download ZIP" feature to download sketches, rather than doing a Git clone, and GitHub appends the branch/tag/commit to the downloaded folder name, which will cause the folder to still not match the sketch name.